+++ CHANGELOG.md
@@ -14,6 +14,8 @@
   unpack/decompress and does not leak temporary TAR files on fail
 * Fixed symlinks scopes for `extractFileToDisk` that could overwrite existing file on disk outside of output dir
 * Fixed CRC64 `XZEncoder` created files created on dart-js failed to decode on native platforms
+* CRC64 for XZDecoder now works both on io and web (dart2-js) (with new `Crc64` class). Old api `isCrc64Supported` 
+  remained unchanged and still returns false and throws with `getCrc64` on dart2-js.
 * Fixed zlib encoder on web
 * Added `CodecsRecognizer` with `ArchiveFormat` to recognize compressed file types with their header bytes
 * Fixed `XZEncoder` (without compression) for large files